+++ examples/dojo/src/widgets/styles/Menu.m.css.d.ts
@@ -1,3 +1,0 @@
-export const root: string;
-export const link: string;
-export const selected: string;