+++ packages/remix/src/build-vite.ts
@@ -1,7 +1,7 @@
 import { readFileSync, promises as fs, statSync, existsSync } from 'fs';
-import { basename, dirname, join, relative, sep } from 'path';
+import { basename, dirname, join, relative } from 'path';
 import { isErrnoException } from '@vercel/error-utils';
-import { nodeFileTrace } from '@vercel/nft';
+import { nodeFileTrace, NodeFileTraceOptions } from '@vercel/nft';
 import {
   BuildResultV2Typical,
   debug,
@@ -13,7 +13,6 @@ import {
   runNpmInstall,
   runPackageJsonScript,
   scanParentDirs,
-  FileBlob,
   FileFsRef,
   EdgeFunction,
   NodejsLambda,
@@ -21,9 +20,10 @@ import {
 import {
   getPathFromRoute,
   getRegExpFromPath,
-  getRemixVersion,
+  getPackageVersion,
   hasScript,
   logNftWarnings,
+  findConfig,
 } from './utils';
 import type { BuildV2, Files, NodeVersion } from '@vercel/build-utils';
 
@@ -37,8 +37,12 @@ const nodeServerSrcPromise = fs.readFile(
   join(DEFAULTS_PATH, 'server-node.mjs'),
   'utf-8'
 );
+const reactRouterServerSrcPromise = fs.readFile(
+  join(DEFAULTS_PATH, 'server-react-router.mjs'),
+  'utf-8'
+);
 
-interface RemixBuildResult {
+interface BuildResultBase {
   buildManifest: {
     serverBundles?: Record<
       string,
@@ -57,14 +61,159 @@ interface RemixBuildResult {
       }
     >;
   };
-  remixConfig: {
-    buildDirectory: string;
-  };
   viteConfig?: {
     build?: {
       assetsDir: string;
     };
   };
+}
+
+interface RemixBuildResult extends BuildResultBase {
+  remixConfig: {
+    buildDirectory: string;
+  };
+}
+
+interface ReactRouterBuildResult extends BuildResultBase {
+  reactRouterConfig: {
+    buildDirectory: string;
+  };
+}
+
+type BuildResult = RemixBuildResult | ReactRouterBuildResult;
+
+interface RenderFunctionOptions {
+  nodeVersion: NodeVersion;
+  entrypointDir: string;
+  rootDir: string;
+  serverBuildPath: string;
+  serverEntryPoint: string | undefined;
+  frameworkVersion: string;
+  config: /*TODO: ResolvedNodeRouteConfig*/ any;
+}
+
+interface FrameworkSettings {
+  primaryPackageName: string;
+  buildCommand: string;
+  buildResultFilePath: string;
+  slug: string;
+  sourceSearchValue: string;
+
+  createRenderFunction: (
+    options: RenderFunctionOptions
+  ) => Promise<EdgeFunction | NodejsLambda>;
+}
+
+const REMIX_FRAMEWORK_SETTINGS: FrameworkSettings = {
+  primaryPackageName: '@remix-run/dev',
+  buildCommand: 'remix build',
+  buildResultFilePath: '.vercel/remix-build-result.json',
+  slug: 'remix',
+  sourceSearchValue: '@remix-run/dev/server-build',
+
+  createRenderFunction({
+    nodeVersion,
+    entrypointDir,
+    rootDir,
+    serverBuildPath,
+    serverEntryPoint,
+    frameworkVersion,
+    config,
+  }: RenderFunctionOptions): Promise<EdgeFunction | NodejsLambda> {
+    if (config.runtime === 'edge') {
+      return createRenderEdgeFunction(
+        entrypointDir,
+        rootDir,
+        serverBuildPath,
+        serverEntryPoint,
+        frameworkVersion,
+        config
+      );
+    }
+
+    return createRenderNodeFunction(
+      nodeVersion,
+      entrypointDir,
+      rootDir,
+      serverBuildPath,
+      serverEntryPoint,
+      frameworkVersion,
+      config
+    );
+  },
+};
+
+const REACT_ROUTER_FRAMEWORK_SETTINGS: FrameworkSettings = {
+  primaryPackageName: 'react-router',
+  buildCommand: 'react-router build',
+  buildResultFilePath: '.vercel/react-router-build-result.json',
+  slug: 'react-router',
+  sourceSearchValue: 'ENTRYPOINT_PLACEHOLDER',
+
+  createRenderFunction({
+    nodeVersion,
+    entrypointDir,
+    rootDir,
+    serverBuildPath,
+    serverEntryPoint,
+    frameworkVersion,
+    config,
+  }: RenderFunctionOptions): Promise<EdgeFunction | NodejsLambda> {
+    return createRenderReactRouterFunction(
+      nodeVersion,
+      entrypointDir,
+      rootDir,
+      serverBuildPath,
+      serverEntryPoint,
+      frameworkVersion,
+      config
+    );
+  },
+};
+
+function determineFrameworkSettings(workPath: string) {
+  const isReactRouter = findConfig(workPath, 'react-router.config', [
+    '.js',
+    '.ts',
+    '.mjs',
+    '.mts',
+  ]);
+
+  if (isReactRouter) {
+    return REACT_ROUTER_FRAMEWORK_SETTINGS;
+  }
+  return REMIX_FRAMEWORK_SETTINGS;
+}
+
+interface HandlerOptions {
+  rootDir: string;
+  serverBuildPath: string;
+  serverEntryPoint?: string;
+  serverSourcePromise: Promise<string>;
+  sourceSearchValue: string;
+}
+
+async function determineHandler({
+  rootDir,
+  serverBuildPath,
+  serverEntryPoint,
+  serverSourcePromise,
+  sourceSearchValue,
+}: HandlerOptions) {
+  let handler = relative(rootDir, serverBuildPath);
+  let handlerPath = join(rootDir, handler);
+  if (!serverEntryPoint) {
+    const baseServerBuildPath = basename(serverBuildPath, '.js');
+    handler = join(dirname(handler), `server-${baseServerBuildPath}.mjs`);
+    handlerPath = join(rootDir, handler);
+
+    const serverSource = await serverSourcePromise;
+    await fs.writeFile(
+      handlerPath,
+      serverSource.replace(sourceSearchValue, `./${baseServerBuildPath}.js`)
+    );
+  }
+  return { handler, handlerPath };
 }
 
 export const build: BuildV2 = async ({
@@ -77,6 +226,8 @@ export const build: BuildV2 = async ({
   const { installCommand, buildCommand } = config;
   const mountpoint = dirname(entrypoint);
   const entrypointFsDirname = join(workPath, mountpoint);
+
+  const frameworkSettings = determineFrameworkSettings(workPath);
 
   // Run "Install Command"
   const nodeVersion = await getNodeVersion(
@@ -122,9 +273,14 @@ export const build: BuildV2 = async ({
     await runNpmInstall(entrypointFsDirname, [], spawnOpts, meta, nodeVersion);
   }
 
-  // Determine the version of Remix based on the `@remix-run/dev`
-  // package version.
-  const remixVersion = await getRemixVersion(entrypointFsDirname, repoRootPath);
+  // Determine the version of framework:
+  //   Remix - use "@remix-run/dev"
+  //   React Router - use "react-router"
+  const frameworkVersion = await getPackageVersion(
+    frameworkSettings.primaryPackageName,
+    entrypointFsDirname,
+    repoRootPath
+  );
 
   // Run "Build Command"
   if (buildCommand) {
@@ -145,21 +301,21 @@ export const build: BuildV2 = async ({
       debug(`Executing "build" script`);
       await runPackageJsonScript(entrypointFsDirname, 'build', spawnOpts);
     } else {
-      await execCommand('remix build', {
+      await execCommand(frameworkSettings.buildCommand, {
         ...spawnOpts,
         cwd: entrypointFsDirname,
       });
     }
   }
 
-  const remixBuildResultPath = join(
+  const buildResultJsonPath = join(
     entrypointFsDirname,
-    '.vercel/remix-build-result.json'
+    frameworkSettings.buildResultFilePath
   );
-  let remixBuildResult: RemixBuildResult | undefined;
+  let buildResult: BuildResult | undefined;
   try {
-    const remixBuildResultContents = readFileSync(remixBuildResultPath, 'utf8');
-    remixBuildResult = JSON.parse(remixBuildResultContents);
+    const buildResultContents = readFileSync(buildResultJsonPath, 'utf8');
+    buildResult = JSON.parse(buildResultContents);
   } catch (err: unknown) {
     if (!isErrnoException(err) || err.code !== 'ENOENT') {
       throw err;
@@ -170,7 +326,7 @@ export const build: BuildV2 = async ({
     const buildDirectory = join(entrypointFsDirname, 'build');
     if (statSync(buildDirectory).isDirectory()) {
       console.warn('WARN: The `vercelPreset()` Preset was not detected.');
-      remixBuildResult = {
+      buildResult = {
         buildManifest: {
           routes: {
             root: {
@@ -195,10 +351,10 @@ export const build: BuildV2 = async ({
       // Detect if a server build exists (won't be the case when `ssr: false`)
       const serverPath = 'build/server/index.js';
       if (existsSync(join(entrypointFsDirname, serverPath))) {
-        remixBuildResult.buildManifest.routeIdToServerBundleId = {
+        buildResult.buildManifest.routeIdToServerBundleId = {
           'routes/_index': '',
         };
-        remixBuildResult.buildManifest.serverBundles = {
+        buildResult.buildManifest.serverBundles = {
           '': {
             id: '',
             file: serverPath,
@@ -209,40 +365,33 @@ export const build: BuildV2 = async ({
     }
   }
 
-  if (!remixBuildResult) {
+  if (!buildResult) {
     throw new Error(
       'Could not determine build output directory. Please configure the `vercelPreset()` Preset from the `@vercel/remix` npm package'
     );
   }
 
-  const { buildManifest, remixConfig, viteConfig } = remixBuildResult;
+  const { buildManifest, viteConfig } = buildResult;
+  const buildDirectory =
+    'remixConfig' in buildResult
+      ? buildResult.remixConfig.buildDirectory
+      : buildResult.reactRouterConfig.buildDirectory;
 
-  const staticDir = join(remixConfig.buildDirectory, 'client');
+  const staticDir = join(buildDirectory, 'client');
   const serverBundles = Object.values(buildManifest.serverBundles ?? {});
 
   const [staticFiles, ...functions] = await Promise.all([
     glob('**', staticDir),
     ...serverBundles.map(bundle => {
-      if (bundle.config.runtime === 'edge') {
-        return createRenderEdgeFunction(
-          entrypointFsDirname,
-          repoRootPath,
-          join(entrypointFsDirname, bundle.file),
-          undefined,
-          remixVersion,
-          bundle.config
-        );
-      }
-
-      return createRenderNodeFunction(
+      return frameworkSettings.createRenderFunction({
         nodeVersion,
-        entrypointFsDirname,
-        repoRootPath,
-        join(entrypointFsDirname, bundle.file),
-        undefined,
-        remixVersion,
-        bundle.config
-      );
+        entrypointDir: entrypointFsDirname,
+        rootDir: repoRootPath,
+        serverBuildPath: join(entrypointFsDirname, bundle.file),
+        serverEntryPoint: undefined,
+        frameworkVersion,
+        config: bundle.config,
+      });
     }),
   ]);
 
@@ -300,8 +449,122 @@ export const build: BuildV2 = async ({
     dest: '/',
   });
 
-  return { routes, output, framework: { version: remixVersion } };
+  return { routes, output, framework: { version: frameworkVersion } };
 };
+
+async function edgeReadFile(fsPath: string) {
+  let source: Buffer | string;
+  try {
+    source = await fs.readFile(fsPath);
+  } catch (err: any) {
+    if (err.code === 'ENOENT' || err.code === 'EISDIR') {
+      return null;
+    }
+    throw err;
+  }
+  if (basename(fsPath) === 'package.json') {
+    // For Edge Functions, patch "main" field to prefer "browser" or "module"
+    const pkgJson = JSON.parse(source.toString());
+
+    for (const prop of ['browser', 'module']) {
+      const val = pkgJson[prop];
+      if (typeof val === 'string') {
+        pkgJson.main = val;
+
+        // Return the modified `package.json` to nft
+        source = JSON.stringify(pkgJson);
+        break;
+      }
+    }
+  }
+  return source;
+}
+
+const EDGE_TRACE_CONDITIONS = [
+  'edge-light',
+  'browser',
+  'module',
+  'import',
+  'require',
+];
+
+const COMMON_NODE_FUNCTION_OPTIONS = {
+  shouldAddHelpers: false,
+  shouldAddSourcemapSupport: false,
+  operationType: 'SSR',
+  supportsResponseStreaming: true,
+} as const;
+
+const COMMON_EDGE_FUNCTION_OPTIONS = { deploymentTarget: 'v8-worker' } as const;
+
+async function createRenderReactRouterFunction(
+  nodeVersion: NodeVersion,
+  entrypointDir: string,
+  rootDir: string,
+  serverBuildPath: string,
+  serverEntryPoint: string | undefined,
+  frameworkVersion: string,
+  config: /*TODO: ResolvedNodeRouteConfig*/ any
+): Promise<EdgeFunction | NodejsLambda> {
+  const isEdgeFunction = config.runtime === 'edge';
+
+  const { handler, handlerPath } = await determineHandler({
+    rootDir,
+    serverBuildPath,
+    serverEntryPoint,
+    serverSourcePromise: reactRouterServerSrcPromise,
+    sourceSearchValue: REACT_ROUTER_FRAMEWORK_SETTINGS.sourceSearchValue,
+  });
+
+  // Trace the handler with `@vercel/nft`
+  let conditions: NodeFileTraceOptions['conditions'];
+  let readFile: NodeFileTraceOptions['readFile'];
+  if (isEdgeFunction) {
+    conditions = EDGE_TRACE_CONDITIONS;
+    readFile = edgeReadFile;
+  }
+  const trace = await nodeFileTrace([handlerPath], {
+    base: rootDir,
+    processCwd: entrypointDir,
+    conditions,
+    readFile,
+  });
+
+  logNftWarnings(trace.warnings, 'react-router');
+
+  const files = await getFilesFromTrace({ fileList: trace.fileList, rootDir });
+
+  let fn: NodejsLambda | EdgeFunction;
+  if (isEdgeFunction) {
+    fn = new EdgeFunction({
+      ...COMMON_EDGE_FUNCTION_OPTIONS,
+      files,
+      entrypoint: handler,
+      regions: config.regions,
+      framework: {
+        slug: REACT_ROUTER_FRAMEWORK_SETTINGS.slug,
+        version: frameworkVersion,
+      },
+    });
+  } else {
+    fn = new NodejsLambda({
+      ...COMMON_NODE_FUNCTION_OPTIONS,
+      files,
+      handler,
+      runtime: nodeVersion.runtime,
+      useWebApi: true,
+      regions: config.regions,
+      memory: config.memory,
+      maxDuration: config.maxDuration,
+      framework: {
+        slug: REACT_ROUTER_FRAMEWORK_SETTINGS.slug,
+        version: frameworkVersion,
+      },
+    });
+  }
+
+  return fn;
+}
 
 async function createRenderNodeFunction(
   nodeVersion: NodeVersion,
@@ -309,28 +572,16 @@ async function createRenderNodeFunction(
   rootDir: string,
   serverBuildPath: string,
   serverEntryPoint: string | undefined,
-  remixVersion: string,
+  frameworkVersion: string,
   config: /*TODO: ResolvedNodeRouteConfig*/ any
 ): Promise<NodejsLambda> {
-  const files: Files = {};
-
-  let handler = relative(rootDir, serverBuildPath);
-  let handlerPath = join(rootDir, handler);
-  if (!serverEntryPoint) {
-    const baseServerBuildPath = basename(serverBuildPath, '.js');
-    handler = join(dirname(handler), `server-${baseServerBuildPath}.mjs`);
-    handlerPath = join(rootDir, handler);
-
-    // Copy the `server-node.mjs` file into the "build" directory
-    const nodeServerSrc = await nodeServerSrcPromise;
-    await fs.writeFile(
-      handlerPath,
-      nodeServerSrc.replace(
-        '@remix-run/dev/server-build',
-        `./${baseServerBuildPath}.js`
-      )
-    );
-  }
+  const { handler, handlerPath } = await determineHandler({
+    rootDir,
+    serverBuildPath,
+    serverEntryPoint,
+    serverSourcePromise: nodeServerSrcPromise,
+    sourceSearchValue: REMIX_FRAMEWORK_SETTINGS.sourceSearchValue,
+  });
 
   // Trace the handler with `@vercel/nft`
   const trace = await nodeFileTrace([handlerPath], {
@@ -340,24 +591,19 @@ async function createRenderNodeFunction(
 
   logNftWarnings(trace.warnings, '@remix-run/node');
 
-  for (const file of trace.fileList) {
-    files[file] = await FileFsRef.fromFsPath({ fsPath: join(rootDir, file) });
-  }
+  const files = await getFilesFromTrace({ fileList: trace.fileList, rootDir });
 
   const fn = new NodejsLambda({
+    ...COMMON_NODE_FUNCTION_OPTIONS,
     files,
     handler,
     runtime: nodeVersion.runtime,
-    shouldAddHelpers: false,
-    shouldAddSourcemapSupport: false,
-    operationType: 'SSR',
-    supportsResponseStreaming: true,
     regions: config.regions,
     memory: config.memory,
     maxDuration: config.maxDuration,
     framework: {
-      slug: 'remix',
-      version: remixVersion,
+      slug: REMIX_FRAMEWORK_SETTINGS.slug,
+      version: frameworkVersion,
     },
   });
 
@@ -369,89 +615,53 @@ async function createRenderEdgeFunction(
   rootDir: string,
   serverBuildPath: string,
   serverEntryPoint: string | undefined,
-  remixVersion: string,
+  frameworkVersion: string,
   config: /* TODO: ResolvedEdgeRouteConfig*/ any
 ): Promise<EdgeFunction> {
-  const files: Files = {};
-
-  let handler = relative(rootDir, serverBuildPath);
-  let handlerPath = join(rootDir, handler);
-  if (!serverEntryPoint) {
-    const baseServerBuildPath = basename(serverBuildPath, '.js');
-    handler = join(dirname(handler), `server-${baseServerBuildPath}.mjs`);
-    handlerPath = join(rootDir, handler);
-
-    // Copy the `server-edge.mjs` file into the "build" directory
-    const edgeServerSrc = await edgeServerSrcPromise;
-    await fs.writeFile(
-      handlerPath,
-      edgeServerSrc.replace(
-        '@remix-run/dev/server-build',
-        `./${baseServerBuildPath}.js`
-      )
-    );
-  }
-
-  let remixRunVercelPkgJson: string | undefined;
+  const { handler, handlerPath } = await determineHandler({
+    rootDir,
+    serverBuildPath,
+    serverEntryPoint,
+    serverSourcePromise: edgeServerSrcPromise,
+    sourceSearchValue: REMIX_FRAMEWORK_SETTINGS.sourceSearchValue,
+  });
 
   // Trace the handler with `@vercel/nft`
   const trace = await nodeFileTrace([handlerPath], {
     base: rootDir,
     processCwd: entrypointDir,
-    conditions: ['edge-light', 'browser', 'module', 'import', 'require'],
-    async readFile(fsPath) {
-      let source: Buffer | string;
-      try {
-        source = await fs.readFile(fsPath);
-      } catch (err: any) {
-        if (err.code === 'ENOENT' || err.code === 'EISDIR') {
-          return null;
-        }
-        throw err;
-      }
-      if (basename(fsPath) === 'package.json') {
-        // For Edge Functions, patch "main" field to prefer "browser" or "module"
-        const pkgJson = JSON.parse(source.toString());
-
-        for (const prop of ['browser', 'module']) {
-          const val = pkgJson[prop];
-          if (typeof val === 'string') {
-            pkgJson.main = val;
-
-            // Return the modified `package.json` to nft
-            source = JSON.stringify(pkgJson);
-            break;
-          }
-        }
-      }
-      return source;
-    },
+    conditions: EDGE_TRACE_CONDITIONS,
+    readFile: edgeReadFile,
   });
 
   logNftWarnings(trace.warnings, '@remix-run/server-runtime');
 
-  for (const file of trace.fileList) {
-    if (
-      remixRunVercelPkgJson &&
-      file.endsWith(`@remix-run${sep}vercel${sep}package.json`)
-    ) {
-      // Use the modified `@remix-run/vercel` package.json which contains "browser" field
-      files[file] = new FileBlob({ data: remixRunVercelPkgJson });
-    } else {
-      files[file] = await FileFsRef.fromFsPath({ fsPath: join(rootDir, file) });
-    }
-  }
+  const files = await getFilesFromTrace({ fileList: trace.fileList, rootDir });
 
   const fn = new EdgeFunction({
+    ...COMMON_EDGE_FUNCTION_OPTIONS,
     files,
-    deploymentTarget: 'v8-worker',
     entrypoint: handler,
     regions: config.regions,
     framework: {
-      slug: 'remix',
-      version: remixVersion,
+      slug: REMIX_FRAMEWORK_SETTINGS.slug,
+      version: frameworkVersion,
     },
   });
 
   return fn;
+}
+
+async function getFilesFromTrace({
+  fileList,
+  rootDir,
+}: {
+  fileList: Set<string>;
+  rootDir: string;
+}) {
+  const files: Files = {};
+  for (const file of fileList) {
+    files[file] = await FileFsRef.fromFsPath({ fsPath: join(rootDir, file) });
+  }
+  return files;
 }