+++ packages/frameworks/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vercel/frameworks
 
+## 3.6.0
+
+### Minor Changes
+
+- Add "react-router" framework entry ([#12932](https://github.com/vercel/vercel/pull/12932))
+
+## 3.5.0
+
+### Minor Changes
+
+- Make vite detection supersede ionic-react ([#12880](https://github.com/vercel/vercel/pull/12880))
+
 ## 3.4.0
 
 ### Minor Changes